Extract UI from onCreate
